Committing changes made for DeskApp version 2.3

+++ DeskApp/src/DeskApp/DataLayer/Eval/dof_blgf_financial_data.cs
@@ -16,21 +16,21 @@ namespace DeskApp.DataLayer.Eval
 
         public int year_id { get; set; }
         public int psgc_code { get; set; }
-        public int? locally_shared_revenues { get; set; }
-        public int? ira_share { get; set; }
-        public int? other_revenues_total { get; set; }
-        public int? other_shares_natl_tax { get; set; }
-        public int? inter_local_transfers { get; set; }
-        public int? extraordinary_receipts { get; set; }
-        public int? total_lgu_income { get; set; }
-        public int? expenditures_gen_public_services { get; set; }
-        public int? expenditures_educ_culture_etc { get; set; }
-        public int? expenditures_health_services { get; set; }
-        public int? expenditures_labor_and_employment { get; set; }
-        public int? expenditures_housing_comm_devt { get; set; }
-        public int? expenditures_social_welfare_services { get; set; }
-        public int? expenditures_economic_services { get; set; }
-        public int? expenditures_other_purposes { get; set; }
+        public float? locally_shared_revenues { get; set; }
+        public float? ira_share { get; set; }
+        public float? other_revenues_total { get; set; }
+        public float? other_shares_natl_tax { get; set; }
+        public float? inter_local_transfers { get; set; }
+        public float? extraordinary_receipts { get; set; }
+        public float? total_lgu_income { get; set; }
+        public float? expenditures_gen_public_services { get; set; }
+        public float? expenditures_educ_culture_etc { get; set; }
+        public float? expenditures_health_services { get; set; }
+        public float? expenditures_labor_and_employment { get; set; }
+        public float? expenditures_housing_comm_devt { get; set; }
+        public float? expenditures_social_welfare_services { get; set; }
+        public float? expenditures_economic_services { get; set; }
+        public float? expenditures_other_purposes { get; set; }
 
     }
